requirements.txt
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
docutils==0.14
idna==2.8
lxml==4.3.2
nltk==3.4
numpy==1.16.2
pex==1.6.3
pkginfo==1.5.0.1
Pygments==2.3.1
readability-lxml==0.7
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.9.1
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.8
strsim==0.0.3
tqdm==4.31.1
twine==1.13.0
urllib3==1.24.1
webencodings==0.5.1
